backend/requirements.txt
# =============================================================================
# Identity Shield - Backend Requirements
# =============================================================================

# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
cryptography==42.0.0

# Cloudinary
cloudinary==1.38.0

# Environment Variables
python-decouple==3.8
python-dotenv==1.0.0

# Utilities
Pillow==10.2.0
pytz==2024.1

# Development Tools
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2
